+++ HISTORY.md
@@ -17,6 +17,7 @@
 ### 2022-03-17
 * blocks: refactored dependencies scan for scripts
 * gui: refactored palette serialization for scripts
+* byob, blocks, gui: refactored blocksLibraryXML()
 
 ### 2022-03-16
 * restored v7.4.0-dev